Fixed color API, cleaned, fixed array impl, added Math::Min, Max, Round and vectorized all math functions, reformed the ThreadPool class. Added FloatInput UI element class

+++ BlazeEngineCore/include/BlazeEngine/Core/Resource/ResourceRef.h
@@ -21,7 +21,7 @@ namespace Blaze
 		bool IsLoaded() const;
 		void WaitToLoad() const;
 
-		uint32 Hash() const;
+		uint64 Hash() const;
 
 		ResourceBase* GetResource() const;
 		template<typename T>
